@@ -1,5 +1,5 @@
 ﻿@page "/"
-@using BouncingBalls.Lab10TODO
+@using BouncingBalls.Lab12TODO
 
 <PageTitle>Index</PageTitle>
 
